requirements.txt
requests>=2.28.1
tomli>=2.0.1
python-dateutil>=2.8.2
boto3>=1.26.0
